// src/app/(calculator)/types/calculation.ts
export interface CalculationResults {
  grossIncome: number;
  employerContributions: number;
  totalCost: number;
  netSalary: number;
  incomeTax: number;
  profitAfterSalary: number;
  corporateTax: number;
  profitAfterTaxes: number;
  annualPension: number;
  annualPensionTax: number;
  dividendAmount: number;
  dividendTax: number;
  netDividend: number;
  remainingCapital: number;
}